+++ myapp/docker-compose.yml
@@ -16,7 +16,7 @@ services:
 
   api:
     build: .
-    command: bash -c "bundle install && bundle exec rails db:create && bundle exec rails db:migrate && rm -f tmp/pids/server.pid && bundle exec rails s -p 3000 -b '0.0.0.0'"
+    command: bash -c "bundle install && yarn install && bundle exec rails db:create && bundle exec rails db:migrate && rm -f tmp/pids/server.pid && bundle exec rails s -p 3000 -b '0.0.0.0'"
     volumes:
       - .:/myapp
       - bundle-install:/usr/local/bundle:delegated
